+++ packages/noodles-cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.3](https://github.com/geallenboy/noodles/compare/noodles-cli@0.0.1...noodles-cli@0.0.3) (2020-04-27)
+
+**Note:** Version bump only for package noodles-cli
+
+
+
+
+
+## [0.0.2](https://github.com/geallenboy/noodles/compare/noodles-cli@0.0.1...noodles-cli@0.0.2) (2020-04-27)
+
+**Note:** Version bump only for package noodles-cli
+
+
+
+
+
 ## 0.0.1 (2020-04-27)
 
 **Note:** Version bump only for package noodles-cli